pyproject.toml
[tool.poetry]
name = "law firm pgvector-rag"
version = "0.1.0"
description = "RAG application with pgvector and TimescaleDB"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/danielgraviet/pgvector_RAG"
repository = "https://github.com/danielgraviet/pgvector_RAG"
keywords = ["rag", "pgvector", "timescaledb", "fastapi", "llm"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"

# Core deps from requirements.txt
aiohappyeyeballs = "2.6.1"
aiohttp = "3.11.16"
aiosignal = "1.3.2"
annotated-types = "0.7.0"
anthropic = "0.49.0"
anyio = "4.9.0"
async-timeout = "5.0.1"
asyncpg = "0.30.0"
attrs = "25.3.0"
certifi = "2025.1.31"
charset-normalizer = "3.4.1"
click = "8.1.8"
colorama = "0.4.6"
distro = "1.9.0"
docstring-parser = "0.16"
exceptiongroup = "1.2.2"
fastapi = "0.115.12"
frozenlist = "1.5.0"
h11 = "0.14.0"
httpcore = "1.0.7"
httpx = "0.28.1"
idna = "3.10"
instructor = "1.7.9"
jinja2 = "3.1.6"
jiter = "0.8.2"
markdown-it-py = "3.0.0"
markupsafe = "3.0.2"
mdurl = "0.1.2"
multidict = "6.4.2"
numpy = "2.0.2"
openai = "1.72.0"
packaging = "24.2"
pandas = "2.2.3"
pgvector = "0.4.0"
propcache = "0.3.1"
psycopg2 = "2.9.10"
psycopg2-binary = "2.9.10"
pydantic = "2.11.3"
pydantic-core = "2.33.1"
pygments = "2.19.1"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.1.0"
pytz = "2025.2"
requests = "2.32.3"
rich = "13.9.4"
shellingham = "1.5.4"
six = "1.17.0"
sniffio = "1.3.1"
sqlalchemy = "2.0.40"
starlette = "0.46.2"
tenacity = "9.1.2"
timescale-vector = "0.0.7"
tomli = "2.2.1"
tqdm = "4.67.1"
typer = "0.15.2"
typing-inspection = "0.4.0"
typing-extensions = "4.13.1"
tzdata = "2025.2"
urllib3 = "2.4.0"
uvicorn = "0.34.1"
yarl = "1.19.0"

# Editable local package
pgvector-rag = { git = "https://github.com/danielgraviet/pgvector_RAG.git", rev = "571ae4cbca00b86a0b96ba3436d160f309977b6d" }

[tool.poetry.group.dev.dependencies]
pytest = "8.3.5"
pytest-mock = "3.14.0"

[tool.poetry.scripts]
pgvector-rag = "app.main:run"  # adjust entrypoint

[[tool.poetry.packages]]
include = "app"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"